+++ src/components/InputContainer/Password/index.jsx
@@ -1,22 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { EXPIRATION_DATE_TYPE } from '../../types';
+import { PASSWORD_TYPE } from '../../types';
 import { isNumberInRange } from '../../InputForm/validation';
-import Input from '..';
 import { uid } from 'react-uid';
 import { findNotCompletedInput } from '../../../utils/util';
+import InputContainer from '..';
 
-function ExpirationDateInput({ expirationDate, cardInputDispatch, inputElementsRef, stateName }) {
-  const onChangeExpirationDate = (e, key) => {
+function PasswordInput({ password, cardInputDispatch, inputElementsRef, stateName }) {
+  const onChangePassword = (e, key) => {
     const {
-      target: { value: date, maxLength },
+      target: { value: password, maxLength },
     } = e;
 
-    if (isNumberInRange(date, maxLength)) {
-      cardInputDispatch({ type: 'CHANGE_EXPIRATION_DATE', payload: { date, key } });
+    if (isNumberInRange(password, maxLength)) {
+      cardInputDispatch({
+        type: 'CHANGE_PASSWORD',
+        payload: { password, key },
+      });
     }
 
-    if (date.length === maxLength) {
+    if (password.length === maxLength) {
       const { current: inputElementsMap } = inputElementsRef;
 
       const {
@@ -29,16 +32,15 @@ function ExpirationDateInput({ expirationDate, cardInputDispatch, inputElementsR
     }
   };
   return (
-    <Input labelTitle="만료일 (01~12의 월 / 년도)" inputSize="w-50">
-      {Object.keys(expirationDate).map(stateKey => (
+    <InputContainer labelTitle="카드 비밀번호" inputSize="w-50">
+      {Object.keys(password).map(stateKey => (
         <input
           key={uid(stateKey)}
           className="input-basic"
           type="text"
-          placeholder={stateKey === 'month' ? 'MM' : 'YY'}
-          value={expirationDate[stateKey]}
-          onChange={e => onChangeExpirationDate(e, stateKey)}
-          maxLength={2}
+          value={password[stateKey]}
+          onChange={e => onChangePassword(e, stateKey)}
+          maxLength={1}
           required
           ref={element => {
             const { current } = inputElementsRef;
@@ -50,14 +52,15 @@ function ExpirationDateInput({ expirationDate, cardInputDispatch, inputElementsR
           }}
         />
       ))}
-    </Input>
+      <div className="inputted-password">*</div>
+      <div className="inputted-password">*</div>
+    </InputContainer>
   );
 }
-
-ExpirationDateInput.propTypes = {
-  expirationDate: EXPIRATION_DATE_TYPE,
+PasswordInput.propTypes = {
+  password: PASSWORD_TYPE,
   cardInputDispatch: PropTypes.func,
   inputElementsRef: PropTypes.object,
   stateName: PropTypes.string,
 };
-export default ExpirationDateInput;
+export default PasswordInput;
