+++ src/components/InputContainer/SecurityCode/index.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { isNumberInRange } from '../../InputForm/validation';
-import Input from '..';
 import { findNotCompletedInput } from '../../../utils/util';
+import { useEffect } from 'react';
+import { useFindNextInput } from '../../../hooks/useNextInputFoucs';
+import InputContainer from '..';
 
 function SecurityCodeInput({ securityCode, cardInputDispatch, inputElementsRef, stateName }) {
   const onChangeSecurityCode = e => {
@@ -31,7 +33,7 @@ function SecurityCodeInput({ securityCode, cardInputDispatch, inputElementsRef, 
   };
 
   return (
-    <Input
+    <InputContainer
       labelTitle="보안코드(CVC/CVV)"
       inputSize="w-25"
       helpText="카드 뒷면 서명란 또는 신용카드 번호 오른쪽 상단에 기재된 3자리 숫자"
@@ -52,7 +54,7 @@ function SecurityCodeInput({ securityCode, cardInputDispatch, inputElementsRef, 
           };
         }}
       />
-    </Input>
+    </InputContainer>
   );
 }
 SecurityCodeInput.propTypes = {
