+++ src/components/InputContainer/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useMouseHover } from '../../hooks/useMouseHover';
 
-function Input({ labelTitle, inputSize = '', helpText, children }) {
+function InputContainer({ labelTitle, inputSize = '', helpText, children }) {
   const [isMouseOver, onMouseOver, onMouseLeave] = useMouseHover();
 
   return (
@@ -23,11 +23,11 @@ function Input({ labelTitle, inputSize = '', helpText, children }) {
   );
 }
 
-Input.propTypes = {
+InputContainer.propTypes = {
   labelTitle: PropTypes.string,
   inputSize: PropTypes.string,
   helpText: PropTypes.string,
   children: PropTypes.node,
 };
 
-export default Input;
+export default InputContainer;
